refactoring semantic kernel connectors databases
