+++ src/Databases/Milvus/src/MilvusVectorDatabase.cs
@@ -1,44 +1,6 @@
-﻿using Microsoft.SemanticKernel.Connectors.Milvus;
+﻿using LangChain.Databases.Connectors;
+using Microsoft.SemanticKernel.Connectors.Milvus;
 
 namespace LangChain.Databases.Milvus;
 
-public class MilvusVectorDatabase(
-    MilvusMemoryStore store)
-    : IVectorDatabase
-{
-    public async Task CreateCollectionAsync(string collectionName, int dimensions, CancellationToken cancellationToken = default)
-    {
-        await store.CreateCollectionAsync(collectionName, cancellationToken).ConfigureAwait(false);
-    }
-
-    public async Task DeleteCollectionAsync(string collectionName, CancellationToken cancellationToken = default)
-    {
-        await store.DeleteCollectionAsync(collectionName, cancellationToken).ConfigureAwait(false);
-    }
-
-    public async Task<IVectorCollection> GetCollectionAsync(string collectionName, CancellationToken cancellationToken = default)
-    {
-        var collections = await ListCollectionsAsync(cancellationToken).ConfigureAwait(false);
-        var collection = collections.FirstOrDefault(x => x == collectionName);
-        return collection != null ? new MilvusVectorCollection(store, collection)
-            : throw new InvalidOperationException("Collection not found");
-    }
-
-    public async Task<IVectorCollection> GetOrCreateCollectionAsync(string collectionName, int dimensions, CancellationToken cancellationToken = default)
-    {
-        if (!await IsCollectionExistsAsync(collectionName, cancellationToken).ConfigureAwait(false))
-            await store.CreateCollectionAsync(collectionName, cancellationToken).ConfigureAwait(false);
-        return new MilvusVectorCollection(store, collectionName);
-    }
-
-    public async Task<bool> IsCollectionExistsAsync(string collectionName, CancellationToken cancellationToken = default)
-    {
-        return await store.DoesCollectionExistAsync(collectionName, cancellationToken).ConfigureAwait(false);
-    }
-
-    public async Task<IReadOnlyList<string>> ListCollectionsAsync(CancellationToken cancellationToken = default)
-    {
-        var collections = store.GetCollectionsAsync(cancellationToken);
-        return await collections.ToListAsync(cancellationToken).ConfigureAwait(false);
-    }
-}
+public class MilvusVectorDatabase(MilvusMemoryStore store) : SemanticKernelMemoryDatabase(store);
