+++ src/Databases/Pinecone/src/LangChain.Databases.Pinecone.csproj
@@ -1,7 +1,8 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
         <TargetFrameworks>net4.6.2;netstandard2.0;net6.0;net8.0</TargetFrameworks>
+		<NoWarn>$(NoWarn);SKEXP0020;SKEXP0001;CS3001</NoWarn>
     </PropertyGroup>
 
     <PropertyGroup Label="NuGet">
@@ -11,6 +12,7 @@
 
     <ItemGroup>
         <PackageReference Include="Microsoft.SemanticKernel.Connectors.Pinecone" />
+		<ProjectReference Include="..\..\Connectors\src\LangChain.Databases.Connectors.csproj" />
     </ItemGroup>
 
 </Project>
