+++ src/Databases/Pinecone/src/PineconeVectorCollection.cs
@@ -1,10 +1,10 @@
 ﻿using LangChain.Databases.Connectors;
-using Microsoft.SemanticKernel.Connectors.DuckDB;
+using Microsoft.SemanticKernel.Connectors.Pinecone;
 
-namespace LangChain.Databases.DuckDb
+namespace LangChain.Databases.Pinecone
 {
-    public class DuckDbVectorCollection(
-        DuckDBMemoryStore store,
+    public class PineconeVectorCollection(
+        PineconeMemoryStore store,
         string name = VectorCollection.DefaultName,
         string? id = null) : SemanticKernelMemoryStoreCollection(store, name, id);
 }
