+++ src/Databases/Pinecone/src/PineconeVectorDatabase.cs
@@ -1,0 +1,7 @@
+﻿using LangChain.Databases.Connectors;
+using Microsoft.SemanticKernel.Connectors.Pinecone;
+
+namespace LangChain.Databases.Pinecone
+{
+    public class MilvusVectorDatabase(PineconeMemoryStore store) : SemanticKernelMemoryDatabase(store);
+}
