+++ src/Databases/Redis/src/RedisVectorCollection.cs
@@ -1,0 +1,10 @@
+﻿using LangChain.Databases.Connectors;
+using Microsoft.SemanticKernel.Connectors.Redis;
+
+namespace LangChain.Databases.Redis
+{
+    public class RedisVectorCollection(
+    RedisMemoryStore store,
+    string name = VectorCollection.DefaultName,
+    string? id = null) : SemanticKernelMemoryStoreCollection(store, name, id);
+}
