+++ src/Databases/Redis/src/RedisVectorDatabase.cs
@@ -1,0 +1,7 @@
+﻿using LangChain.Databases.Connectors;
+using Microsoft.SemanticKernel.Connectors.Redis;
+
+namespace LangChain.Databases.Redis
+{
+    public class RedisVectorDatabase(RedisMemoryStore store) : SemanticKernelMemoryDatabase(store);
+}
