+++ src/Databases/Weaviate/src/WeaviateVectorCollection.cs
@@ -1,0 +1,10 @@
+﻿using LangChain.Databases.Connectors;
+using Microsoft.SemanticKernel.Connectors.Weaviate;
+
+namespace LangChain.Databases.Weaviate
+{
+    public class WeaviateVectorCollection(
+    WeaviateMemoryStore store,
+    string name = VectorCollection.DefaultName,
+    string? id = null) : SemanticKernelMemoryStoreCollection(store, name, id);
+}
